simplify even more

+++ src/database/TrainData.py
@@ -9,4 +9,4 @@ class TrainData(SplitDataSource):
         return None
 
     def row_to(self):
-        return int(math.floor(self._split_row))
+        return math.floor(self._split_row)
